third_party/WebKit/LayoutTests/fast/multicol/client-rects-empty-element-boundary.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>body { margin:0; }</style>
<div style="columns:4; column-gap:10px; column-fill:auto; width:430px; height:100px;">
    <div style="height:100px;"></div>
    <div id="empty" style="position:relative;"></div>
    <div style="height:100px;"></div>
</div>
<script>
    test(function() {
        var rects = document.getElementById("empty").getClientRects();
        assert_equals(rects.length, 1);
        assert_equals(rects[0].left, 110);
        assert_equals(rects[0].top, 0);
        assert_equals(rects[0].right, 210);
        assert_equals(rects[0].bottom, 0);
    }, "Zero-height element at column boundary");
</script>
